+++ src/registry/defaultRegistry.ts
@@ -18,6 +18,13 @@ import { HiddenField } from '../components/fields/HiddenField'
 import { ArrayField } from '../components/fields/ArrayField'
 import { ObjectField } from '../components/fields/ObjectField'
 import { CustomField } from '../components/fields/CustomField'
+import { PinField } from '../components/fields/PinField'
+import { RatingField } from '../components/fields/RatingField'
+import { EditableField } from '../components/fields/EditableField'
+import { SegmentedField } from '../components/fields/SegmentedField'
+import { TagsField } from '../components/fields/TagsField'
+import { CheckboxCardField } from '../components/fields/CheckboxCardField'
+import { RadioCardField } from '../components/fields/RadioCardField'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any -- registry must accept any field component props
 const entries: [FieldType | string, ComponentType<any>][] = [
@@ -34,6 +41,9 @@ const entries: [FieldType | string, ComponentType<any>][] = [
   ['multiselect', MultiSelectField],
   ['radio', RadioField],
   ['checkboxgroup', CheckboxGroupField],
+  ['checkboxcard', CheckboxCardField],
+  ['radiocard', RadioCardField],
+  ['segmented', SegmentedField],
   ['boolean', BooleanField],
   ['switch', BooleanField],
   ['checkbox', BooleanField],
@@ -43,6 +53,10 @@ const entries: [FieldType | string, ComponentType<any>][] = [
   ['file', FileField],
   ['slider', SliderField],
   ['color', ColorField],
+  ['pin', PinField],
+  ['rating', RatingField],
+  ['editable', EditableField],
+  ['tags', TagsField],
   ['hidden', HiddenField],
   ['array', ArrayField],
   ['object', ObjectField],